requirements.txt
fastapi==0.114.2
uvicorn==0.30.6
pydantic==2.9.2
requests==2.32.3
python-dotenv==1.0.1
trafilatura==1.9.0
beautifulsoup4==4.12.3
streamlit==1.38.0
google-generativeai==0.7.2
psycopg2-binary==2.9.9
reportlab==4.2.2
notion-client==2.2.1
google-search-results==2.4.2
pandas==2.2.2
